requirements.txt
python-dotenv
tk
pyrebase
qrcode
pillow